More fix for x64 Also fix dbeditor from 8572 commit
git-svn-id: http://svn.miranda-ng.org/main/trunk@8584 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Clist_mw/src/CLUIFrames/cluiframes.cpp
@@ -933,7 +933,7 @@ INT_PTR CLUIFramesSetFrameOptions(WPARAM wParam, LPARAM lParam)
 
 		SendMessage(Frames[pos].TitleBar.hwndTip,TTM_ACTIVATE,(WPARAM)Frames[pos].TitleBar.ShowTitleBarTip,0);
 
-		style = (int)GetWindowLongPtr(Frames[pos].hWnd,GWL_STYLE);
+		style = GetWindowLongPtr(Frames[pos].hWnd,GWL_STYLE);
 		style |= WS_BORDER;
 		if (flag&F_NOBORDER) {style &= (~WS_BORDER);}
 			SetWindowLongPtr(Frames[pos].hWnd,GWL_STYLE,(LONG_PTR)style);
